Fix #73: Only remove files from gh-pages when there are files to remove.

+++ src/main/groovy/org/ajoberstar/gradle/git/ghpages/GithubPagesPlugin.groovy
@@ -58,7 +58,9 @@ class GithubPagesPlugin implements Plugin<Project> {
 			def filesList = extension.workingDir.list({ dir, name ->
 				return !name.equals('.git')
 			})
-			repo.remove(patterns: filesList)
+			if (filesList) {
+				repo.remove(patterns: filesList)
+			}
 			project.copy {
 				with extension.pages
 				into repo.repository.rootDir
